+++ tests-corpus/valid-bare-schema/union-aliased-lax-void/out.gen.d.ts
@@ -7,10 +7,6 @@ export type u64 = bigint
 
 export type Void = undefined | null
 
-export function readVoid(bc: bare.ByteCursor): Void
-
-export function writeVoid(bc: bare.ByteCursor, x: Void): void
-
 export type UnsignedInt =
     | { readonly tag: 0; readonly val: u8 }
     | { readonly tag: 1; readonly val: u16 }